将 hazel 工程从 dll 换成 static lib , 不知道为啥 Core.h 的改动会导致一个 CreateApplication 的link error，所以先没有去掉 HAZEL_API; 新增 renderContext

+++ Hazel-learn/src/Hazel/ImGui/ImGuiLayer.cpp
@@ -82,7 +82,7 @@ namespace Hazel
 		// Rendering
 		ImGuiIO& io = ImGui::GetIO();
 		auto& app = Application::Get();
-		io.DisplaySize = ImVec2(app.GetWindow().Width(), app.GetWindow().Height());
+		io.DisplaySize = ImVec2((float)app.GetWindow().Width(), (float)app.GetWindow().Height());
 		
         ImGui::Render();
         ImGui_ImplOpenGL3_RenderDrawData(ImGui::GetDrawData());
