+++ Hazel-learn/src/Platform/Windows/WindowsInput.cpp
@@ -38,6 +38,6 @@ namespace Hazel
 		auto* const window = static_cast<GLFWwindow*>(Application::Get().GetWindow().GetNativeWindow());
 		double x, y;
 		glfwGetCursorPos(window, &x, &y);
-		return {x, y};
+		return {(float)x, (float)y};
 	}
 }
